+++ PCIT_core/internal/llvm_interface/include/IRBuilder.h
@@ -76,7 +76,7 @@ namespace pcit::llvmint{
 
 			auto createUnreachable() -> void;
 			
-			auto createBranch(const BasicBlock& block) -> void;
+			auto createBranch(const BasicBlock& block) -> BranchInst;
 
 			auto createCondBranch(const Value& cond, const BasicBlock& then_block, const BasicBlock& else_block)
 				-> void;