+++ PCIT_core/internal/llvm_interface/include/class_impls/native_ptr_decls.h
@@ -52,6 +52,7 @@ namespace llvm{
 	class Constant;
 	class ConstantInt;
 	class ReturnInst;
+	class BranchInst;
 	class CallInst;
 	class LoadInst;
 	class GlobalVariable;